Switch from uint64_t to int
While it is prettier in some sense to explicitly state that we intend it to only be nonnegative,
it only provides more issues than anything else. So, let's just completely stick with using int
rather than uint64_t

+++ src/sylvan/package_mgr.h
@@ -66,16 +66,16 @@ public:
   inline sylvan::Bdd leaf_false()
   { return sylvan::Bdd::bddZero(); }
 
-  inline sylvan::Bdd ithvar(size_t label)
+  inline sylvan::Bdd ithvar(int label)
   { return sylvan::Bdd::bddVar(label); }
 
-  inline sylvan::Bdd nithvar(size_t label)
+  inline sylvan::Bdd nithvar(int label)
   { return ~sylvan::Bdd::bddVar(label); }
 
   inline sylvan::Bdd ite(const sylvan::Bdd &f, const sylvan::Bdd &g, const sylvan::Bdd &h)
   { return f.Ite(g,h); }
 
-  inline sylvan::Bdd exists(const sylvan::Bdd &b, size_t label)
+  inline sylvan::Bdd exists(const sylvan::Bdd &b, int label)
   { return b.ExistAbstract(sylvan::Bdd::bddVar(label)); }
 
   inline uint64_t nodecount(const bdd_t &b)
